Module fasthash::t1ha::t1ha1

t1ha1 = 64-bit, BASELINE FAST PORTABLE HASH:

Runs faster on 64-bit platforms in other cases may runs slowly.

Portable and stable, returns same 64-bit result on all architectures and CPUs.

Unfortunately it fails the "strict avalanche criteria", see test results at https://github.com/demerphq/smhasher.

This flaw is insignificant for the t1ha1() purposes and imperceptible from a practical point of view. However, nowadays this issue has resolved in the next t1ha2(), that was initially planned to providing a bit more quality.

Re-exports

pub use self::Hasher64Le as Hasher64;

pub use self::Hash64Le as Hash64;

Structs

Hash64Be
T1Hash 64-bit hash functions for 64-bit big-endian platforms.

Hash64Le
T1Hash 64-bit hash functions for 64-bit little-endian platforms.

Hasher64Be
An implementation of std::hash::Hasher.

Hasher64Le
An implementation of std::hash::Hasher.