Module fasthash::t1ha::t1ha2

t1ha2 = 64 and 128-bit, SLIGHTLY MORE ATTENTION FOR QUALITY AND STRENGTH.

The recommended version of "Fast Positive Hash" with good quality for checksum, hash tables and fingerprinting.

Portable and extremely efficiency on modern 64-bit CPUs. Designed for 64-bit little-endian platforms, in other cases will runs slowly.

Great quality of hashing and still faster than other non-t1ha hashes. Provides streaming mode and 128-bit result.

Structs

Hash64AtOnce
The at-once variant with 64-bit result

Hash128AtOnce
The at-once variant with 64-bit result

Hasher64
An implementation of std::hash::Hasher.

Hasher128
An implementation of std::hash::Hasher and fasthash::HasherExt.