Trait fasthash::HasherExt

pub trait HasherExt: Hasher {
    fn finish_ext(&self) -> u128;
}

A trait which represents the ability to hash an arbitrary stream of bytes.

Required methods

fn finish_ext(&self) -> u128

Completes a round of hashing, producing the output hash generated.

Loading content...