Trait fasthash::HasherExt

pub trait HasherExt: Hasher {
    fn finish_ext(&self) -> u128;

    fn write_u128(&mut self, i: u128) { ... }
    fn write_i128(&mut self, i: i128) { ... }
}

A trait which represents the ability to hash an arbitrary stream of bytes.

Required Methods

fn finish_ext(&self) -> u128

Completes a round of hashing, producing the output hash generated.

Provided Methods

fn write_u128(&mut self, i: u128)

Writes a single u128 into this hasher.

fn write_i128(&mut self, i: i128)

Writes a single i128 into this hasher.